refactor(package): 프로젝트 이름 Image -> image로 수정 및 모듈로 추가
