--- Image/build.gradle
@@ -1,13 +1,0 @@
-dependencies {
-	implementation project(':common')
-	testImplementation(testFixtures('common'))
-
-	implementation 'org.springframework.boot:spring-boot-starter-web'
-	implementation 'org.springframework.boot:spring-boot-starter-data-jpa'
-	// AWS SDK v2 - S3 업로드 용
-	implementation("software.amazon.awssdk:s3:2.20.100")
-	// 이미지 MIME 타입 처리용
-	implementation 'org.apache.tika:tika-core:2.8.0'
-
-	runtimeOnly 'com.mysql:mysql-connector-j'
-}
