--- Image/src/main/java/rice_monkey/Image/ImageApplication.java
@@ -1,13 +1,0 @@
-package rice_monkey.Image;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class ImageApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(ImageApplication.class, args);
-	}
-
-}
