--- Image/src/main/java/rice_monkey/Image/dto/ImageResponse.java
@@ -1,15 +1,0 @@
-package rice_monkey.Image.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-
-@Getter
-@AllArgsConstructor
-public class ImageResponse {
-
-    private Long id;
-
-    private String url;
-
-    private String fileName;
-}
