--- Image/src/main/java/rice_monkey/Image/dto/ImageUploadResponse.java
@@ -1,12 +1,0 @@
-package rice_monkey.Image.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-
-@Getter
-@AllArgsConstructor
-public class ImageUploadResponse {
-    private Long imageId;
-    private String url;
-}
-
