--- Image/src/main/java/rice_monkey/Image/entity/Image.java
@@ -1,26 +1,0 @@
-package rice_monkey.Image.entity;
-
-import jakarta.persistence.*;
-import lombok.*;
-
-import java.time.LocalDateTime;
-
-@Entity
-@Getter
-@Setter
-@NoArgsConstructor
-@AllArgsConstructor
-@Builder
-public class Image {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    private String url;
-    private String fileName;
-    private String contentType;
-    private Long size;
-    private LocalDateTime createdAt;
-    private Boolean isDeleted;
-}
