--- Image/src/test/java/rice_monkey/Image/ImageApplicationTests.java
@@ -1,13 +1,0 @@
-package rice_monkey.Image;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class ImageApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
