+++ image/build.gradle
@@ -1,0 +1,14 @@
+dependencies {
+	implementation project(':common')
+	testImplementation(testFixtures(project(':common')))
+
+	implementation 'org.springframework.boot:spring-boot-starter-web'
+	implementation 'org.springframework.boot:spring-boot-starter-data-jpa'
+	implementation 'org.flywaydb:flyway-mysql'
+	// AWS SDK v2 - S3 업로드 용
+	implementation("software.amazon.awssdk:s3:2.20.100")
+	// 이미지 MIME 타입 처리용
+	implementation 'org.apache.tika:tika-core:2.8.0'
+
+	runtimeOnly 'com.mysql:mysql-connector-j'
+}
