+++ image/gradle/wrapper/gradle-wrapper.properties
