+++ image/src/main/java/rice_monkey/image/ImageApplication.java
@@ -1,0 +1,13 @@
+package rice_monkey.image;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class ImageApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(ImageApplication.class, args);
+	}
+
+}
