+++ image/src/main/java/rice_monkey/image/dto/ImageResponse.java
@@ -1,0 +1,15 @@
+package rice_monkey.image.dto;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+
+@Getter
+@AllArgsConstructor
+public class ImageResponse {
+
+    private Long id;
+
+    private String url;
+
+    private String fileName;
+}
