+++ image/src/main/java/rice_monkey/image/dto/ImageUploadResponse.java
@@ -1,0 +1,12 @@
+package rice_monkey.image.dto;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+
+@Getter
+@AllArgsConstructor
+public class ImageUploadResponse {
+    private Long imageId;
+    private String url;
+}
+
