+++ image/src/main/java/rice_monkey/image/entity/Image.java
@@ -1,0 +1,26 @@
+package rice_monkey.image.entity;
+
+import jakarta.persistence.*;
+import lombok.*;
+
+import java.time.LocalDateTime;
+
+@Entity
+@Getter
+@Setter
+@NoArgsConstructor
+@AllArgsConstructor
+@Builder
+public class Image {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+
+    private String url;
+    private String fileName;
+    private String contentType;
+    private Long size;
+    private LocalDateTime createdAt;
+    private Boolean isDeleted;
+}
