+++ image/src/main/java/rice_monkey/image/repository/ImageRepository.java
@@ -1,0 +1,9 @@
+package rice_monkey.image.repository;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+import rice_monkey.image.entity.Image;
+
+@Repository
+public interface ImageRepository extends JpaRepository<Image, Long> {}
+
