+++ image/src/main/java/rice_monkey/image/util/DummyUploader.java
@@ -1,0 +1,14 @@
+package rice_monkey.image.util;
+
+import org.springframework.stereotype.Service;
+import org.springframework.web.multipart.MultipartFile;
+
+@Service
+public class DummyUploader implements Uploader {
+    @Override
+    public String upload(MultipartFile file) {
+        System.out.println("📦 Dummy upload: " + file.getOriginalFilename());
+        return "https://dummy.local/" + file.getOriginalFilename();
+    }
+}
+
