+++ image/src/main/java/rice_monkey/image/util/S3Uploader.java
@@ -1,0 +1,105 @@
+package rice_monkey.image.util;
+
+import lombok.RequiredArgsConstructor;
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.stereotype.Service;
+import org.springframework.web.multipart.MultipartFile;
+import software.amazon.awssdk.core.sync.RequestBody;
+import software.amazon.awssdk.services.s3.S3Client;
+import software.amazon.awssdk.services.s3.model.*;
+
+import java.io.IOException;
+import java.io.InputStream;
+import java.util.ArrayList;
+import java.util.List;
+import java.util.UUID;
+
+@Service
+@RequiredArgsConstructor
+public class S3Uploader implements Uploader {
+
+    private final S3Client s3Client;
+
+    @Value("${cloud.aws.s3.bucket}")
+    private String bucket;
+
+    @Value("${cloud.aws.s3.url}")
+    private String s3BaseUrl;
+
+    private static final long PART_SIZE = 5 * 1024 * 1024; //5MB씩 잘라서 남은 데이터들 다 받기
+
+    @Override
+    public String upload(MultipartFile file) throws IOException {
+        String key = UUID.randomUUID() + "_" + file.getOriginalFilename();
+
+        // 1. initiate multipart upload
+        CreateMultipartUploadRequest createRequest = CreateMultipartUploadRequest.builder()
+                .bucket(bucket)
+                .key(key)
+                .contentType(file.getContentType())
+                .build();
+
+        CreateMultipartUploadResponse createResponse = s3Client.createMultipartUpload(createRequest);
+        String uploadId = createResponse.uploadId();
+
+        List<CompletedPart> completedParts = new ArrayList<>();
+        try (InputStream inputStream = file.getInputStream()) {
+            byte[] buffer = new byte[(int) PART_SIZE];
+            int bytesRead;
+            int partNumber = 1;
+            while ((bytesRead = inputStream.read(buffer)) != -1) {
+                UploadPartRequest uploadRequest = UploadPartRequest.builder()
+                        .bucket(bucket)
+                        .key(key)
+                        .uploadId(uploadId)
+                        .partNumber(partNumber)
+                        .contentLength((long) bytesRead)
+                        .build();
+
+                UploadPartResponse uploadResponse = s3Client.uploadPart(uploadRequest,
+                        RequestBody.fromBytes(copyOf(buffer, bytesRead)));
+
+                completedParts.add(CompletedPart.builder()
+                        .partNumber(partNumber)
+                        .eTag(uploadResponse.eTag())
+                        .build());
+
+                partNumber++;
+            }
+
+            // 3. complete upload
+            CompletedMultipartUpload completedMultipartUpload = CompletedMultipartUpload.builder()
+                    .parts(completedParts)
+                    .build();
+
+            CompleteMultipartUploadRequest completeRequest = CompleteMultipartUploadRequest.builder()
+                    .bucket(bucket)
+                    .key(key)
+                    .uploadId(uploadId)
+                    .multipartUpload(completedMultipartUpload)
+                    .build();
+
+            s3Client.completeMultipartUpload(completeRequest);
+
+            return s3BaseUrl + "/" + key;
+
+        } catch (Exception e) {
+            // 4. abort upload on error
+            AbortMultipartUploadRequest abortRequest = AbortMultipartUploadRequest.builder()
+                    .bucket(bucket)
+                    .key(key)
+                    .uploadId(uploadId)
+                    .build();
+            s3Client.abortMultipartUpload(abortRequest);
+            throw new IOException("Multipart upload failed", e);
+        }
+    }
+
+    private byte[] copyOf(byte[] source, int length) {
+        byte[] copy = new byte[length];
+        System.arraycopy(source, 0, copy, 0, length);
+        return copy;
+    }
+}
+
+
