+++ image/src/main/java/rice_monkey/image/util/Uploader.java
@@ -1,0 +1,9 @@
+package rice_monkey.image.util;
+
+import org.springframework.web.multipart.MultipartFile;
+
+import java.io.IOException;
+
+public interface Uploader {
+    String upload(MultipartFile file) throws IOException;
+}
