+++ image/src/main/resources/application.properties
@@ -1,0 +1,14 @@
+spring.application.name=image
+
+# MySQL ??
+spring.datasource.url=jdbc:mysql://localhost:3306/issue_tracker?serverTimezone=Asia/Seoul
+spring.datasource.username=JOHNNIE
+spring.datasource.password=
+spring.datasource.driver-class-name=com.mysql.cj.jdbc.Driver
+
+# AWS S3 ??
+cloud.aws.region.static=ap-northeast-2
+cloud.aws.s3.bucket=my-issue-image-bucket
+cloud.aws.s3.url=https://my-issue-image-bucket.s3.ap-northeast-2.amazonaws.com
+cloud.aws.credentials.access-key=dummy
+cloud.aws.credentials.secret-key=dummy
