+++ image/src/test/java/rice_monkey/image/ImageApplicationTests.java
@@ -1,0 +1,13 @@
+package rice_monkey.image;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class ImageApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
